database: Move some functions to operator module.
